workflows: new jumpstart workflow
Merge the `build` (`pull_request`) and `publish` (`workflow_run`) workflows into a
single `jumpstart` (`pull_request_target`) workflow.

The main issue that this solves is that it makes it much easier to use the
actually built commit ID (the merge commit for pull requests, or the head
commit for the main branch) as an identifier for the resulting cached object.
This solves the inconsistency issues we've been seeing.

We are now building arbitrary code in `pull_request_target`.  We manage this
risk in two ways:

  - the build process (including the checking out of the code) takes place
    as a normal user, inside of a container.  It should be impossible to break
    out.

  - even if a container escape is possible, the VM running the build job has a
    severely restricted GITHUB_TOKEN and no access to secrets

Building inside of a container also gets us a newer version of nodejs, which
slightly increases performance.

We drop the old workflows which used to be responsible for this process.